# open-api/pkg/model/aether-2.0.0-openapi3.yaml
...
components:
  parameters:
    target:
      content:
        text/plain; charset=utf-8:
          schema:
            type: string
      description: target (target in onos-config)
      in: path
      name: target
      required: true
  requestBodies:
...
    RequestBody_Enterprises:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Enterprises'
...
  schemas:
...
    AdditionalPropertyTarget:
      description: Optionally specify a target other than the default (only on PATCH
        method)
      properties:
        target:
          description: an override of the target (target)
          title: target
          type: string
      title: AdditionalPropertyTarget
      type: object
    AdditionalPropertyUnchanged:
      description: To optionally omit 'required' properties, add them to 'unchanged'
        list
      properties:
        unchanged:
          description: A comma seperated list of unchanged mandatory attribute names
          title: unchanged
          type: string
      title: AdditionalPropertyUnchanged
      type: object
...
    Enterprises:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyTarget'
      description: The top level enterprises container
      properties:
        enterprise:
          $ref: '#/components/schemas/Enterprises_Enterprise_List'
      title: Enterprises
      type: object
    Enterprises_Enterprise:
      description: List of enterprises (single)
      properties:
        application:
          $ref: '#/components/schemas/Enterprises_Enterprise_Application_List'
        connectivity-service:
          $ref: '#/components/schemas/Enterprises_Enterprise_Connectivity-service_List'
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        enterprise-id:
          description: ID for this enterprise.
          maxLength: 63
          minLength: 1
          pattern: '[a-z]([a-z0-9-]?[a-z0-9])*'
          title: enterprise-id
          type: string
        site:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_List'
        template:
          $ref: '#/components/schemas/Enterprises_Enterprise_Template_List'
        traffic-class:
          $ref: '#/components/schemas/Enterprises_Enterprise_Traffic-class_List'
      required:
      - enterprise-id
      title: Enterprises_Enterprise
      type: object
      x-list-multiple: true
    Enterprises_Enterprise_Application:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: List of applications (single)
      properties:
        address:
          description: Address of this application. Either a hostname, an IP, or a
            subnet.
          title: address
          type: string
        application-id:
          description: ID for this application.
          maxLength: 63
          minLength: 1
          pattern: '[a-z]([a-z0-9-]?[a-z0-9])*'
          title: application-id
          type: string
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        endpoint:
          $ref: '#/components/schemas/Enterprises_Enterprise_Application_Endpoint_List'
      required:
      - address
      - application-id
      title: Enterprises_Enterprise_Application
      type: object
      x-list-multiple: true
...
    Enterprises_Enterprise_Application_Endpoint_List:
      description: list for endpoint (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Application_Endpoint'
      type: array
      uniqueItems: true
      x-keys:
      - endpoint-id
      x-list-multiple: true
...
    Enterprises_Enterprise_Application_List:
      description: List of applications (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Application'
      type: array
      uniqueItems: true
      x-keys:
      - application-id
      x-list-multiple: true
    Enterprises_Enterprise_Connectivity-service:
      description: The list for connectivity-service (single)
      properties:
        connectivity-service:
          description: Link to connectivity services where configuration should be
            pushed for this enterprise's devices
          title: connectivity-service
          type: string
          x-leafref: /cs:connectivity-services/cs:connectivity-service/cs:connectivity-service-id
        enabled:
          default: true
          description: Allow or disallow pushes to this connectivity service
          title: enabled
          type: boolean
      required:
      - connectivity-service
      title: Enterprises_Enterprise_Connectivity-service
      type: object
      x-list-multiple: true
    Enterprises_Enterprise_Connectivity-service_List:
      description: The list for connectivity-service (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Connectivity-service'
      type: array
      uniqueItems: true
      x-keys:
      - connectivity-service
      x-list-multiple: true
    Enterprises_Enterprise_List:
      description: List of enterprises (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise'
      type: array
      uniqueItems: true
      x-keys:
      - enterprise-id
      x-list-multiple: true
    Enterprises_Enterprise_Site:
      description: List of site (single)
      properties:
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        device:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Device_List'
        device-group:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Device-group_List'
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        imsi-definition:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Imsi-definition'
        ip-domain:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Ip-domain_List'
        monitoring:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Monitoring'
        sim-card:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Sim-card_List'
        site-id:
          description: ID for this site.
          maxLength: 63
          minLength: 1
          pattern: '[a-z]([a-z0-9-]?[a-z0-9])*'
          title: site-id
          type: string
        slice:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Slice_List'
        small-cell:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Small-cell_List'
        upf:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Upf_List'
      required:
      - site-id
      title: Enterprises_Enterprise_Site
      type: object
      x-list-multiple: true
      x-must:
      - ErrorMessage:
          Name: a device group cannot be used in more than one slice in a site
        Name: count(ent:slice[set-contains(following-sibling::ent:slice/ent:device-group/@ent:device-group,
          ent:device-group/@ent:device-group)]/@ent:slice-id) = 0
...
    Enterprises_Enterprise_Site_Device-group_List:
      description: List of device groups (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site_Device-group'
      type: array
      uniqueItems: true
      x-keys:
      - device-group-id
      x-list-multiple: true
...
    Enterprises_Enterprise_Site_Device_List:
      description: List of devices (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site_Device'
      type: array
      uniqueItems: true
      x-keys:
      - device-id
      x-list-multiple: true
    Enterprises_Enterprise_Site_Imsi-definition:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: container for imsi-defination
      properties:
        enterprise:
          description: enterprise-specific identifier
          format: int32
          maximum: 999
          minimum: 0
          title: enterprise
          type: integer
        format:
          description: IMSI format specifier, describes how fields are packed into
            an IMSI. Must be exactly 15 characters long. For example, CCCNNNEEESSSSSS.
          maxLength: 15
          minLength: 15
          title: format
          type: string
        mcc:
          description: mobile country code
          pattern: '[0-9]{3}'
          title: mcc
          type: string
        mnc:
          description: mobile network code
          pattern: '[0-9]{2,3}'
          title: mnc
          type: string
      required:
      - enterprise
      - format
      - mcc
      - mnc
      title: Enterprises_Enterprise_Site_Imsi-definition
      type: object
...
    Enterprises_Enterprise_Site_Ip-domain_List:
      description: List of ip domains (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site_Ip-domain'
      type: array
      uniqueItems: true
      x-keys:
      - ip-domain-id
      x-list-multiple: true
    Enterprises_Enterprise_Site_List:
      description: List of site (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site'
      type: array
      uniqueItems: true
      x-keys:
      - site-id
      x-list-multiple: true
    Enterprises_Enterprise_Site_Monitoring:
      description: container for monitoring
      properties:
        edge-cluster-prometheus-url:
          description: URL of edge cluster prometheus
          title: edge-cluster-prometheus-url
          type: string
        edge-device:
          $ref: '#/components/schemas/Enterprises_Enterprise_Site_Monitoring_Edge-device_List'
        edge-monitoring-prometheus-url:
          description: URL of monitoring prometheus
          title: edge-monitoring-prometheus-url
          type: string
      title: Enterprises_Enterprise_Site_Monitoring
      type: object
...
    Enterprises_Enterprise_Site_Sim-card_List:
      description: List of sim cards (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site_Sim-card'
      type: array
      uniqueItems: true
      x-keys:
      - sim-id
      x-list-multiple: true
...
    Enterprises_Enterprise_Site_Slice_List:
      description: List of Slices (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Site_Slice'
      type: array
      uniqueItems: true
      x-keys:
      - slice-id
      x-list-multiple: true
...
    Enterprises_Enterprise_Template:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: List of slice templates (single)
      properties:
        default-behavior:
          description: Default behavior if no filter rules match
          maxLength: 20
          pattern: DENY-ALL|ALLOW-ALL|ALLOW-PUBLIC
          title: default-behavior
          type: string
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        mbr:
          $ref: '#/components/schemas/Enterprises_Enterprise_Template_Mbr'
        sd:
          description: Slice differentiator
          format: int32
          maximum: 16777215
          minimum: 0
          title: sd
          type: integer
        sst:
          description: Slice/Service type
          maximum: 255
          minimum: 1
          title: sst
          type: integer
        template-id:
          description: ID for this slice template.
          maxLength: 63
          minLength: 1
          pattern: '[a-z]([a-z0-9-]?[a-z0-9])*'
          title: template-id
          type: string
      required:
      - default-behavior
      - template-id
      title: Enterprises_Enterprise_Template
      type: object
      x-list-multiple: true
    Enterprises_Enterprise_Template_List:
      description: List of slice templates (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Template'
      type: array
      uniqueItems: true
      x-keys:
      - template-id
      x-list-multiple: true
...
    Enterprises_Enterprise_Traffic-class:
      description: List of traffic class (single)
      properties:
        arp:
          description: Allocation and Retention Priority. 1 is the highest. 15 is
            the lowest.
          maximum: 15
          minimum: 1
          title: arp
          type: integer
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        pdb:
          description: PDB
          maximum: 1000
          minimum: 0
          title: pdb
          type: integer
        pelr:
          description: PELR exponent
          maximum: 10
          minimum: 0
          title: pelr
          type: integer
        qci:
          description: QOS Class Identifier
          maximum: 32
          minimum: 1
          title: qci
          type: integer
        traffic-class-id:
          description: ID for this traffic class.
          maxLength: 63
          minLength: 1
          pattern: '[a-z]([a-z0-9-]?[a-z0-9])*'
          title: traffic-class-id
          type: string
      required:
      - traffic-class-id
      title: Enterprises_Enterprise_Traffic-class
      type: object
      x-list-multiple: true
    Enterprises_Enterprise_Traffic-class_List:
      description: List of traffic class (list)
      items:
        $ref: '#/components/schemas/Enterprises_Enterprise_Traffic-class'
      type: array
      uniqueItems: true
      x-keys:
      - traffic-class-id
      x-list-multiple: true
info:
  contact:
    email: info@opennetworking.org
    name: Open Networking Foundation
    url: https://opennetworking.org
  description: OpenAPI 3 specification is generated from aether onos-config model
    plugin
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: aether-2.0.x
  version: 2.0.x
openapi: 3.0.0
paths:
...
  /aether/v2.0.x/{target}/enterprises:
    delete:
      operationId: deleteEnterprises
      responses:
        "200":
          description: DELETE 200 OK
      summary: DELETE /enterprises
    description: The top level enterprises container
    get:
      operationId: getEnterprises
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprises'
          description: GET OK 200
      summary: GET /enterprises Container
      tags:
      - ""
      - Container
    parameters:
    - $ref: '#/components/parameters/target'
    post:
      operationId: postEnterprises
      requestBody:
        $ref: '#/components/requestBodies/RequestBody_Enterprises'
      responses:
        "201":
          description: created
      summary: POST /enterprises
...